Auth fully functional
