+++ .expo/packager-info.json
@@ -1,0 +1,3 @@
+{
+  "devToolsPort": 19002
+}
